large sweeping cleanup.

+++ core-services/src/main/java/org/springsource/examples/sawt/services/jdbc/Main.java
@@ -20,12 +20,12 @@ public class Main {
 
         Customer customer = customerService.createCustomer("Josh", "Long");
 
-        Customer retrievedCustomer = customerService.getCustomerById(customer.getId().longValue());
+        Customer retrievedCustomer = customerService.getCustomerById(customer.getId());
 
         log.info(String.format("customer.id (%s) == retrievedCustomer.id (%s)?  %s",
                 customer.getId(), retrievedCustomer.getId(), customer.getId().equals(retrievedCustomer.getId())));
 
-        Customer updatedCustomer = customerService.updateCustomer(customer.getId().longValue(), "Jdbc", "Lover");
+        Customer updatedCustomer = customerService.updateCustomer(customer.getId() , "Jdbc", "Lover");
         log.info(String.format("updated customer's firstName: %s", updatedCustomer.getFirstName()));
 
     }
