+++ core-services/src/main/java/org/springsource/examples/sawt/services/jpa/Main.java
@@ -17,12 +17,12 @@ public class Main {
 
         Customer customer = customerService.createCustomer("Jpa", "Lover");
 
-        Customer retrievedCustomer = customerService.getCustomerById(customer.getId().longValue());
+        Customer retrievedCustomer = customerService.getCustomerById(customer.getId ());
 
         log.info(String.format("customer.id (%s) == retreivedCustomer.id (%s)?  %s",
                 customer.getId(), retrievedCustomer.getId(), customer.getId().equals(retrievedCustomer.getId())));
 
-        Customer updatedCustomer = customerService.updateCustomer(customer.getId().longValue(), "JPA", "Lover");
+        Customer updatedCustomer = customerService.updateCustomer(customer.getId() , "JPA", "Lover");
         log.info(String.format("updated customer's firstName: %s", updatedCustomer.getFirstName()));
